+++ BasicModels/Board.cs
@@ -20,7 +20,7 @@ namespace battleship
         [DllImport("kernel32.dll")]
         static extern bool SetConsoleMode(IntPtr hConsoleHandle, uint dwMode);
 
-        private static readonly int BOARD_DIMENSION = 11;
+        public static readonly int BOARD_DIMENSION = 11;
         private static readonly string[] ROW_HEADERS = { "A", "B", "C", "D", "E", "F", "G", "H", "I", "J" };
         private static readonly string[] COLUMN_HEADERS = { "1", "2", "3", "4", "5", "6", "7", "8", "9", "10" };
         public static int ROW_OFFSET = 12;
@@ -42,7 +42,7 @@ namespace battleship
         {
             GameBoard = new string[BOARD_DIMENSION, BOARD_DIMENSION];
             FillBoard();
-            Lives = 5;
+            Lives = 17;
         }
 
         private void FillBoard()
@@ -120,14 +120,84 @@ namespace battleship
             }
         }
 
+        public bool AddComputerBoatToBoard(Boat boat, int row, int column)
+        {
+            if (boat.Orientation == Orientation.Horizontal && 
+                ValidComputerHorizontalBoatLocation(boat, row, column))
+            {
+                AddComputerHorizontalBoat(boat, row, column);
+                return true;
+            }
+            else if (boat.Orientation == Orientation.Vertical && 
+                ValidComputerVerticalBoatLocation(boat, row, column))
+            {
+                AddComputerVerticalBoat(boat, row, column);
+                return true;
+            }
+            else
+            {
+                return false;
+            }
+        }
+
+        private bool ValidComputerHorizontalBoatLocation(Boat boat, int row, int column)
+        {
+            for (int i = 0; i < boat.BoatLives.Length; i++)
+            {
+                if (GameBoard[row, column + i] == BOAT_SPACE ||
+                    GameBoard[row, column + i] == BOAT_SPACE_RIGHT_EDGE)
+                {
+                    return false;
+                }
+            }
+            return true;
+        }
+
+        public void AddComputerHorizontalBoat(Boat boat, int row, int column)
+        {
+            for (int i = 0; i < boat.BoatLives.Length; i++)
+            {
+                if (GameBoard[row, column + i] == EMPTY_SPACE)
+                    GameBoard[row, column + i] = BOAT_SPACE;
+                else
+                    GameBoard[row, column + i] = BOAT_SPACE_RIGHT_EDGE;
+            }
+        }
+
+        private bool ValidComputerVerticalBoatLocation(Boat boat, int row, int column)
+        {
+            for (int i = 0; i < boat.BoatLives.Length; i++)
+            {
+                if (GameBoard[row + i, column] == BOAT_SPACE ||
+                    GameBoard[row + i, column] == BOAT_SPACE_RIGHT_EDGE)
+                {
+                    return false;
+                }
+            }
+            return true;
+        }
+
+        public void AddComputerVerticalBoat(Boat boat, int row, int column)
+        {
+            for (int i = 0; i < boat.BoatLives.Length; i++)
+            {
+                if (GameBoard[row + i, column] == EMPTY_SPACE)
+                    GameBoard[row + i, column] = BOAT_SPACE;
+                else
+                    GameBoard[row + i, column] = BOAT_SPACE_RIGHT_EDGE;
+            }
+        }
+
         public bool AddBoatToBoard(Boat boat, int row, int column)
         {
-            if (boat.Orientation == Orientation.Horizontal && ValidHorizontalBoatLocation(boat, row, column))
+            if (boat.Orientation == Orientation.Horizontal && 
+                ValidHorizontalBoatLocation(boat, row, column))
             {
                 AddHorizontalBoat(boat, row, 0, column, 0);
                 return true;
             }
-            else if (boat.Orientation == Orientation.Vertical && ValidVerticalBoatLocation(boat, row, column))
+            else if (boat.Orientation == Orientation.Vertical && 
+                ValidVerticalBoatLocation(boat, row, column))
             {
                 AddVerticalBoat(boat, row, 0, column, 0);
                 return true;
@@ -153,6 +223,19 @@ namespace battleship
             return true;
         }
 
+        public void AddHorizontalBoat(Boat boat, int row, int rowOffset, int column, int columnOffset)
+        { 
+            int boardRow = row - ROW_OFFSET + 1;
+            int boardColumn = TranslateConsoleGridToGameBoard(column);
+            for (int i = 0; i < boat.BoatLives.Length; i++)
+            { 
+                if (GameBoard[boardRow, boardColumn + i] == EMPTY_SPACE)
+                    GameBoard[boardRow, boardColumn + i] = BOAT_SPACE;
+                else
+                    GameBoard[boardRow, boardColumn + i] = BOAT_SPACE_RIGHT_EDGE;
+            }
+        }
+
         private bool ValidVerticalBoatLocation(Boat boat, int row, int column)
         {
             int boardRow = row - ROW_OFFSET + 1;
@@ -168,18 +251,6 @@ namespace battleship
             return true;
         }
 
-        public void AddHorizontalBoat(Boat boat, int row, int rowOffset, int column, int columnOffset)
-        { 
-            int boardRow = row - ROW_OFFSET + 1;
-            int boardColumn = TranslateConsoleGridToGameBoard(column);
-            for (int i = 0; i < boat.BoatLives.Length; i++)
-            { 
-                if (GameBoard[boardRow, boardColumn + i] == EMPTY_SPACE)
-                    GameBoard[boardRow, boardColumn + i] = BOAT_SPACE;
-                else
-                    GameBoard[boardRow, boardColumn + i] = BOAT_SPACE_RIGHT_EDGE;
-            }
-        }
 
         public void AddVerticalBoat(Boat boat, int row, int rowOffset, int column, int columnOffset)
         {