+++ PlaceBoatBehaviors/ComputerBoatPlacer.cs
@@ -8,7 +8,32 @@ namespace battleship.PlaceBoatBehaviors
     {
         public void PlaceBoat(Board board)
         {
-            Console.WriteLine("Computer placing a boat");
+            Random randomNumber = new Random();
+            int numberOfBoats = Enum.GetValues(typeof(BoatType)).Length;
+            Boat boat;
+            int boatLength, row, column, orientation;
+
+            for (int i = 0; i < numberOfBoats; i++)
+            {
+                boat = SimpleBoatFactory.INSTANCE.CreateBoat((BoatType)i);
+                boatLength = boat.BoatLives.Length;
+                do
+                {
+                    orientation = randomNumber.Next(0, 2);
+                    boat.Orientation = (Orientation)orientation;
+
+                    if (boat.Orientation == Orientation.Horizontal)
+                    {
+                        row = randomNumber.Next(1, Board.BOARD_DIMENSION);
+                        column = randomNumber.Next(1, Board.BOARD_DIMENSION + 1 - boatLength);
+                    }
+                    else
+                    {
+                        row = randomNumber.Next(1, Board.BOARD_DIMENSION + 1 - boatLength);
+                        column = randomNumber.Next(1, Board.BOARD_DIMENSION);
+                    }
+                } while (!board.AddComputerBoatToBoard(boat, row, column));
+            }
         }
     }
 }